+++ src/app/components/sprints/sprint/sprint.component.html
@@ -1,0 +1,1 @@
+<p>sprint works!</p>
