+++ src/app/components/sprints/sprints.module.ts
@@ -1,0 +1,21 @@
+import { SprintComponent } from './sprint/sprint.component';
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { AppRoutingModule } from 'src/app/app-routing.module';
+import { SharedModule } from '../shared/shared.module';
+import { HttpClientInMemoryWebApiModule } from 'angular-in-memory-web-api';
+import { HttpClientModule } from '@angular/common/http';
+import { InMemoryDataService } from '../../services/in-memory-data.service';
+
+@NgModule({
+  declarations: [SprintComponent],
+  imports: [
+    CommonModule,
+    SharedModule,
+    AppRoutingModule,
+    HttpClientModule,
+    HttpClientInMemoryWebApiModule.forRoot(InMemoryDataService),
+  ],
+  exports: [SprintComponent],
+})
+export class SprintsModule {}
