--- src/app/services/in-memory/data-sprint.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { DataSprintService } from './data-sprint.service';
-
-describe('DataSprintService', () => {
-  let service: DataSprintService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(DataSprintService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
